fix: pack wire structs; add Reply_gpio to ReplyAny; double→float for speed/accel fields
- Add __attribute__((packed)) to all Message/Reply structs with explicit
  padding fields to keep doubles/floats naturally aligned
- Add Reply_gpio to union ReplyAny (was missing)
- Remove stale 'not currently used' comment from Message_joint_enable
- Change max_velocity, max_accel in Message/Reply_joint_config from
  double to float (source values never needed double precision)
- Change Reply_spindle_speed.speed from double to float (source is
  already float; reduces struct from 24 to 16 wire bytes)
- Fix core0.c unpack_joint_abs_pos to copy doubles out of packed struct
  before taking addresses, avoiding unaligned pointer UB on Cortex-M0+
- Update all tests to use aligned32(sizeof(...)) for buffer lengths and
  size assertions, exposing a latent bug where sizeof == aligned32(sizeof)
  only by coincidence before packing

Wire format note: aligned32(sizeof) is unchanged for all structs except
Reply_spindle_speed (24→16 bytes). Firmware reflash and driver recompile
required.

+++ src/driver/hal_rp2040_eth.c
@@ -367,10 +367,10 @@ bool configure_joint(
     skeleton_t *data
 ) {
     bool pack_success = true;
-    double max_velocity_ticks =
-      (*data->joint_max_velocity[joint]) * (*data->joint_scale[joint]);
-    double max_accel_ticks =
-      (*data->joint_max_accel[joint]) * (*data->joint_scale[joint]);
+    float max_velocity_ticks =
+      (float)((*data->joint_max_velocity[joint]) * (*data->joint_scale[joint]));
+    float max_accel_ticks =
+      (float)((*data->joint_max_accel[joint]) * (*data->joint_scale[joint]));
     /* Send if anything changed, or if no reply has arrived yet (last_joint_config
      * only updates in unpack_joint_config on receipt of REPLY_JOINT_CONFIG, so a
      * lost packet leaves the diff intact and causes an automatic retry). */
